Moved some generic functionality to kieker-common
