+++ kieker-monitoring/src/kieker/monitoring/writer/serializer/AbstractContainerFormatSerializer.java
@@ -21,6 +21,7 @@ import java.util.Collection;
 
 import kieker.common.configuration.Configuration;
 import kieker.common.record.IMonitoringRecord;
+import kieker.common.util.dataformat.FormatIdentifier;
 
 /**
  * Abstract serializer for the Kieker container format for monitoring records. This container
@@ -32,8 +33,8 @@ import kieker.common.record.IMonitoringRecord;
  */
 public abstract class AbstractContainerFormatSerializer extends AbstractMonitoringRecordSerializer {
 
-	/** Container format identifier. Reads "KIKA" in ASCII encoding. */
-	public static final int CONTAINER_IDENTIFIER = 0x4B494B41;
+	/** Container format identifier. */
+	public static final int CONTAINER_IDENTIFIER = FormatIdentifier.CONTAINER_FORMAT.getIdentifierValue();
 
 	private static final int HEADER_SIZE = 8;
